+++ lib/article_json/export/amp/elements/image.rb
@@ -3,17 +3,7 @@ module ArticleJSON
     module AMP
       module Elements
         class Image < Base
-          include ArticleJSON::Export::Common::HTML::Elements::Shared::Caption
-          include ArticleJSON::Export::Common::HTML::Elements::Shared::Float
-
-          # Generate the `<figure>` node containing the image and caption
-          # @return [Nokogiri::HTML::NodeSet]
-          def export
-            create_element(:figure, node_opts) do |figure|
-              figure.add_child(image_node)
-              figure.add_child(caption_node(:figcaption))
-            end
-          end
+          include ArticleJSON::Export::Common::HTML::Elements::Image
 
           private
 
@@ -32,12 +22,6 @@ module ArticleJSON
 
           def default_height
             '480'
-          end
-
-          # @return [Hash]
-          def node_opts
-            return if floating_class.nil?
-            { class: floating_class }
           end
         end
       end
