+++ lib/article_json/export/html/elements/image.rb
@@ -3,30 +3,7 @@ module ArticleJSON
     module HTML
       module Elements
         class Image < Base
-          include ArticleJSON::Export::Common::HTML::Elements::Shared::Caption
-          include ArticleJSON::Export::Common::HTML::Elements::Shared::Float
-
-          # Generate the `<figure>` node containing the image and caption
-          # @return [Nokogiri::XML::Element]
-          def export
-            create_element(:figure, node_opts) do |figure|
-              figure.add_child(image_node)
-              figure.add_child(caption_node(:figcaption))
-            end
-          end
-
-          private
-
-          # @return [Nokogiri::XML::NodeSet]
-          def image_node
-            create_element(:img, src: @element.source_url)
-          end
-
-          # @return [Hash]
-          def node_opts
-            return if floating_class.nil?
-            { class: floating_class }
-          end
+          include ArticleJSON::Export::Common::HTML::Elements::Image
         end
       end
     end
